fixing id problem and getRestaurant as manager

+++ foodApp/src/BuildingBlocks/Explorer.BuildingBlocks.Core/Domain/Entity.cs
@@ -2,7 +2,7 @@
 
 public abstract class Entity
 {
-    public long Id { get; protected set; }
+    public int Id { get; protected set; }
 
     public override bool Equals(object? obj)
     {
