+++ foodApp/src/BuildingBlocks/Explorer.BuildingBlocks.Core/UseCases/ICrudRepository.cs
@@ -5,8 +5,8 @@ namespace Explorer.BuildingBlocks.Core.UseCases;
 public interface ICrudRepository<TEntity> where TEntity : Entity
 {
     PagedResult<TEntity> GetPaged(int page, int pageSize);
-    TEntity Get(long id);
+    TEntity Get(int id);
     TEntity Create(TEntity entity);
     TEntity Update(TEntity entity);
-    void Delete(long id);
+    void Delete(int id);
 }
