+++ foodApp/src/Explorer.API/Controllers/RestaurantController.cs
@@ -36,7 +36,7 @@ namespace Explorer.API.Controllers
         // Add Worker to a Restaurant
         [HttpPost("{restaurantId}/workers")]
         [Authorize(Roles = "Manager")]
-        public async Task<ActionResult> AddWorkerToRestaurant(long restaurantId, [FromBody] UserDto workerDto)
+        public async Task<ActionResult> AddWorkerToRestaurant(int restaurantId, [FromBody] UserDto workerDto)
         {
             var result = await _restaurantService.AddWorkerToRestaurantAsync(restaurantId, workerDto);
             return CreateResponse(result);
@@ -45,7 +45,7 @@ namespace Explorer.API.Controllers
         // Add Food to a Restaurant
         [HttpPost("{restaurantId}/foods")]
         [Authorize(Roles = "Manager")]
-        public async Task<ActionResult> AddFoodToRestaurant(long restaurantId, [FromBody] FoodDto foodDto)
+        public async Task<ActionResult> AddFoodToRestaurant(int restaurantId, [FromBody] FoodDto foodDto)
         {
             foodDto.RestaurantId = restaurantId;
             var result = await _restaurantService.AddFoodToRestaurantAsync(foodDto);
