+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.API/Dtos/RatingReportDto.cs
@@ -4,9 +4,9 @@ namespace Explorer.Stakeholders.API.Dtos
 {
     public class RatingReportDto
     {
-        public long Id { get; set; }                // ID prijave
-        public long OrderId { get; set; }           // ID porudžbine (da ne vraćamo ceo Order objekat)
-        public long ManagerId { get; set; }         // ID menadžera koji je obrađuje
+        public int Id { get; set; }                // ID prijave
+        public int OrderId { get; set; }           // ID porudžbine (da ne vraćamo ceo Order objekat)
+        public int ManagerId { get; set; }         // ID menadžera koji je obrađuje
         public string Comment { get; set; }         // Komentar menadžera ili gosta
         public string Status { get; set; }          // Pending, Approved, Rejected
         public DateTime CreatedAt { get; set; }     // Datum i vreme kada je prijava napravljena
