+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.API/Public/IRestaurantApplicationService.cs
@@ -7,6 +7,6 @@ namespace Explorer.Stakeholders.API.Public
         Task<RestaurantApplicationDto> SubmitApplicationAsync(RestaurantApplicationDto dto);
         Task<List<RestaurantApplicationDto>> GetAllApplicationsAsync();
         Task<List<RestaurantApplicationDto>> GetPendingApplicationsAsync();
-        Task<RestaurantApplicationDto> ProcessApplicationAsync(long applicationId, ProcessApplicationDto decision);
+        Task<RestaurantApplicationDto> ProcessApplicationAsync(int applicationId, ProcessApplicationDto decision);
     }
 }
