+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.API/Public/IRestaurantService.cs
@@ -12,7 +12,7 @@ namespace Explorer.Stakeholders.API.Public
     {
         Task<Result<IEnumerable<RestaurantDto>>> GetAllRestaurantsAsync();
         Task<RestaurantDto> AddRestaurantAsync(RestaurantDto dto);
-        Task<Result> AddWorkerToRestaurantAsync(long restaurantId, UserDto workerDto);
+        Task<Result> AddWorkerToRestaurantAsync(int restaurantId, UserDto workerDto);
         Task<Result> AddFoodToRestaurantAsync(FoodDto foodDto);
