+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/Food.cs
@@ -14,9 +14,9 @@ namespace Explorer.Stakeholders.Core.Domain
         public string Description { get; private set; }
         public string ImageUrl { get; private set; }
 
-        public long RestaurantId { get; private set; }  
+        public int RestaurantId { get; private set; }
 
-        public Food(string name, decimal price, string description, string imageUrl, long restaurantId)
+        public Food(string name, decimal price, string description, string imageUrl, int restaurantId)
         {
             Name = name;
             Price = price;
