+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/Person.cs
@@ -5,12 +5,12 @@ namespace Explorer.Stakeholders.Core.Domain;
 
 public class Person : Entity
 {
-    public long UserId { get; init; }
+    public int UserId { get; init; }
     public string Name { get; init; }
     public string Surname { get; init; }
     public string Email { get; init; }
 
-    public Person(long userId, string name, string surname, string email)
+    public Person(int userId, string name, string surname, string email)
     {
         UserId = userId;
         Name = name;
