+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/RepositoryInterfaces/IFoodRepository.cs
@@ -8,7 +8,7 @@ namespace Explorer.Stakeholders.Core.Domain.RepositoryInterfaces
 {
     public interface IFoodRepository
     {
-        Task<IEnumerable<Food>> GetAllFoodByRestaurantAsync(long restaurantId);
-        Task<List<Food>> GetByIdsAsync(IEnumerable<long> ids);
+        Task<IEnumerable<Food>> GetAllFoodByRestaurantAsync(int restaurantId);
+        Task<List<Food>> GetByIdsAsync(IEnumerable<int> ids);
     }
 }
