+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/RepositoryInterfaces/IRestaurantApplicationRepository.cs
@@ -3,7 +3,7 @@ namespace Explorer.Stakeholders.Core.Domain.RepositoryInterfaces
     public interface IRestaurantApplicationRepository
     {
         Task<RestaurantApplication> CreateAsync(RestaurantApplication application);
-        Task<RestaurantApplication?> GetByIdAsync(long id);
+        Task<RestaurantApplication?> GetByIdAsync(int id);
         Task<List<RestaurantApplication>> GetAllAsync();
         Task<List<RestaurantApplication>> GetByStatusAsync(ApplicationStatus status);
         Task<RestaurantApplication> UpdateAsync(RestaurantApplication application);
