+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/Domain/RepositoryInterfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ public interface IUserRepository
     bool Exists(string username);
     User? GetActiveByName(string username);
     User Create(User user);
-    long GetPersonId(long userId);
+    int GetPersonId(int userId);
     Task<User> CreateAsync(User user);
-    Task<User?> GetByIdAsync(long id);
+    Task<User?> GetByIdAsync(int id);
 }
