+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/UseCases/AuthenticationService.cs
@@ -25,7 +25,7 @@ public class AuthenticationService : IAuthenticationService
         var user = _userRepository.GetActiveByName(credentials.Username);
         if (user == null || credentials.Password != user.Password) return Result.Fail(FailureCode.NotFound);
 
-        long personId;
+        int personId;
         try
         {
             personId = _userRepository.GetPersonId(user.Id);
