+++ foodApp/src/Modules/Stakeholders/Explorer.Stakeholders.Core/UseCases/ITokenGenerator.cs
@@ -6,5 +6,5 @@ namespace Explorer.Stakeholders.Core.UseCases;
 
 public interface ITokenGenerator
 {
-    Result<AuthenticationTokensDto> GenerateAccessToken(User user, long personId);
+    Result<AuthenticationTokensDto> GenerateAccessToken(User user, int personId);
 }
